Resolve issues with spawn on latest Boost (v1.88) (#124)
Since Boost 1.88 the default completion token, provided in
boost::asio::spawn's third argument, has changed to
boost::asio::deferred, which results with coroutines not spawning at
all. The common way to fix that is to replace default argument with
boost::asio::detached, but this has a negative effect of not being able
to catch any exception thrown inside a thread. This change addresses the
issue by introducing a wrapper for boost::asio::spawn with custom
completion token, which rethrows any exception that may be thrown inside
a coroutine.

Tested: Unit Test Validation with Boost v.1.88

---------

Signed-off-by: Ewelina Walkusz <ewelinax.walkusz@intel.com>